requirements.txt
Django
djangorestframework
python-telegram-bot
telebot
